Adding BaseStore
