+++ spider_feeder/store/scrapinghub_collection.py
@@ -3,8 +3,10 @@ import os
 from scrapinghub import ScrapinghubClient
 from scrapinghub.client.utils import parse_job_key
 
+from .base_store import BaseStore
 
-class ScrapinghubCollectionStore:
+
+class ScrapinghubCollectionStore(BaseStore):
     '''Store class abstracting `ScrapinghubClient`.
     No configuration for authentication is set. `ScrapinghubClient` loads from environment variables.
     For more information, please refer to https://python-scrapinghub.readthedocs.io/en/latest/client/apidocs.html#scrapinghub.client.ScrapinghubClient.
@@ -13,7 +15,7 @@ class ScrapinghubCollectionStore:
     '''
 
     def __init__(self, input_uri, settings):
-        self._input_field = settings.get('SPIDERFEEDER_INPUT_FIELD')
+        super().__init__(settings)
         client = ScrapinghubClient()
 
         jobkey = parse_job_key(os.environ['SHUB_JOBKEY'])
@@ -22,14 +24,6 @@ class ScrapinghubCollectionStore:
         collection_name = input_uri.replace('collections://', '')
         self._store = project.collections.get_store(collection_name)
 
-    def __iter__(self):
+    def read_input_items(self):
         for item in self._store.iter():
-            data = item['value']
-            if self._input_field:
-                if not isinstance(data, dict):
-                    raise TypeError('Collection data is expected to be a dict when SPIDERFEEDER_INPUT_FIELD is set.')  # noqa
-
-                yield (data[self._input_field], data)
-            else:
-                # TODO: Return dict {} instead of None
-                yield (data, None)
+            yield item['value']
